scraper/obot_scraper/requirements.txt
scrapy
bs4
google-cloud-firestore
google-cloud-aiplatform
langchain-text-splitters

pypdf[crypto]
python-docx
python-pptx
openpyxl